# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Local server"
        },
        {
            "url": "https://foodie-app.azurewebsites.net",
            "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "foodie"
            }
        },
...
    },
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/post": {
            "post": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Create a new post",
                "description": "Create a new post",
                "operationId": "createPost",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "description": "Content",
                                        "example": "test"
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "description": "Privacy",
                                        "example": "public"
                                    },
                                    "photos": {
                                        "type": "array",
                                        "description": "Photos",
                                        "items": {
                                            "type": "string",
                                            "format": "binary",
                                            "description": "Photo"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "description": "Title",
                                            "example": "test"
                                        },
                                        "content": {
                                            "type": "string",
                                            "description": "Content",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/{username}/posts": {
            "get": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts by a user",
                "description": "Get all posts by a user",
                "operationId": "getPosts",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "description": {
                                                "type": "string",
                                                "description": "Content",
                                                "example": "test"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/{post_id}": {
            "patch": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Update a post",
                "description": "Update a post",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "description": "Content",
                                        "example": "test"
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "description": "Privacy",
                                        "example": "public"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "description": "Title",
                                            "example": "test"
                                        },
                                        "content": {
                                            "type": "string",
                                            "description": "Content",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete" : {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Delete a post",
                "description": "Delete a post",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Message",
                                            "example": "Post deleted"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get" : {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get a post",
                "description": "Get a post",
                "operationId": "getPost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "description": "Title",
                                            "example": "test"
                                        },
                                        "content": {
                                            "type": "string",
                                            "description": "Content",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }

            }
        },
        "/api/v1/like/post/{post_id}" : {
            "post" : {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Like"
                ],
                "summary": "Like a post",
                "description": "Like a post",
                "operationId": "likePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Message",
                                            "example": "Post liked"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/post/likes/{post_id}" : {
            "get" : {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Like"
                ],
                "summary": "Get likes of a post",
                "description": "Get likes of a post",
                "operationId": "getPostLikes",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "likes": {
                                            "type": "array",
                                            "description": "Likes",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }

            }
        },
...
    }
}